First pass fix for license listing
