+++ src/cargo-deny/list.rs
@@ -1,21 +1,6 @@
 use crate::common::ValidConfig;
 use anyhow::{Context as _, Error};
-use cargo_deny::{Kid, PathBuf, diag::Files, licenses};
-use nu_ansi_term::Color;
-use serde::Serialize;
-
-#[derive(clap::ValueEnum, Copy, Clone, Debug)]
-pub enum Layout {
-    Crate,
-    License,
-}
-
-#[derive(clap::ValueEnum, Copy, Clone, Debug)]
-pub enum OutputFormat {
-    Human,
-    Json,
-    Tsv,
-}
+use cargo_deny::{PathBuf, diag::Files, licenses};
 
 #[derive(clap::Parser, Debug)]
 pub struct Args {
@@ -40,10 +25,10 @@ pub struct Args {
     threshold: f32,
     /// The format of the output
     #[arg(short, long, default_value = "human", value_enum)]
-    format: OutputFormat,
+    format: licenses::OutputFormat,
     /// The layout for the output, does not apply to TSV
     #[arg(short, long, default_value = "license", value_enum)]
-    layout: Layout,
+    layout: licenses::Layout,
 }
 
 pub fn cmd(
@@ -51,13 +36,12 @@ pub fn cmd(
     args: Args,
     krate_ctx: crate::common::KrateContext,
 ) -> Result<(), Error> {
-    use licenses::LicenseInfo;
-    use std::{collections::BTreeMap, fmt::Write};
-
     let cfg_path = krate_ctx.get_config_path(args.config.as_deref())?;
 
     let mut files = Files::new();
-    let ValidConfig { graph, .. } = ValidConfig::load(
+    let ValidConfig {
+        graph, licenses, ..
+    } = ValidConfig::load(
         cfg_path,
         krate_ctx.get_local_exceptions_path(),
         &mut files,
@@ -85,265 +69,17 @@ pub fn cmd(
 
     let mut files = Files::new();
 
-    let summary = gatherer.gather(&krates, &mut files, None);
+    let summary = gatherer.gather(&krates, &mut files, &licenses);
 
-    use std::borrow::Cow;
+    let colorize = crate::common::should_colorize(log_ctx.color, std::io::stdout());
 
-    #[derive(Ord, PartialOrd, PartialEq, Eq)]
-    struct SerKid<'k>(Cow<'k, Kid>);
-
-    impl serde::Serialize for SerKid<'_> {
-        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: serde::Serializer,
-        {
-            serializer.serialize_str(&format!(
-                "{} {} {}",
-                self.0.name(),
-                self.0.version(),
-                self.0.source()
-            ))
-        }
-    }
-
-    impl SerKid<'_> {
-        fn parts(&self) -> (&str, &str) {
-            (self.0.name(), self.0.version())
-        }
-    }
-
-    #[derive(Serialize)]
-    struct Crate {
-        licenses: Vec<String>,
-    }
-
-    #[derive(Serialize)]
-    struct LicenseLayout<'k> {
-        licenses: Vec<(String, Vec<SerKid<'k>>)>,
-        unlicensed: Vec<SerKid<'k>>,
-    }
-
-    struct CrateLayout<'k> {
-        crates: BTreeMap<SerKid<'k>, Crate>,
-    }
-
-    impl<'k> CrateLayout<'k> {
-        fn search(&self, id: &SerKid<'k>) -> &Crate {
-            self.crates.get(id).expect("unable to find crate")
-        }
-    }
-
-    fn borrow(kid: &Kid) -> SerKid<'_> {
-        SerKid(Cow::Borrowed(kid))
-    }
-
-    let mut crate_layout = CrateLayout {
-        crates: BTreeMap::new(),
-    };
-
-    let mut license_layout = LicenseLayout {
-        licenses: Vec::with_capacity(20),
-        unlicensed: Vec::new(),
-    };
-
-    {
-        let licenses = &mut license_layout.licenses;
-        let unlicensed = &mut license_layout.unlicensed;
-
-        for krate_lic_nfo in summary.nfos {
-            let mut cur = Crate {
-                licenses: Vec::with_capacity(2),
-            };
-
-            match krate_lic_nfo.lic_info {
-                LicenseInfo::SpdxExpression { expr, .. } => {
-                    for req in expr.requirements() {
-                        let s = req.req.to_string();
-
-                        if cur.licenses.contains(&s) {
-                            continue;
-                        }
-
-                        match licenses.binary_search_by(|(r, _)| r.cmp(&s)) {
-                            Ok(i) => licenses[i].1.push(borrow(&krate_lic_nfo.krate.id)),
-                            Err(i) => {
-                                let mut v = Vec::with_capacity(20);
-                                v.push(borrow(&krate_lic_nfo.krate.id));
-                                licenses.insert(i, (s.clone(), v));
-                            }
-                        }
-                        cur.licenses.push(s);
-                    }
-                }
-                LicenseInfo::Unlicensed => {
-                    unlicensed.push(borrow(&krate_lic_nfo.krate.id));
-                }
-            }
-
-            crate_layout
-                .crates
-                .insert(SerKid(Cow::Owned(krate_lic_nfo.krate.id.clone())), cur);
-        }
-    }
-
-    fn write_pid(out: &mut String, pid: &SerKid<'_>) -> Result<(), Error> {
-        let (name, version) = pid.parts();
-        Ok(write!(out, "{name}@{version}")?)
-    }
-
-    match args.format {
-        OutputFormat::Human => {
-            let mut output = String::with_capacity(4 * 1024);
-            let color = crate::common::should_colorize(log_ctx.color, std::io::stdout());
-
-            match args.layout {
-                Layout::License => {
-                    for (license, krates) in license_layout.licenses {
-                        if color {
-                            write!(
-                                output,
-                                "{} ({}): ",
-                                Color::Cyan.paint(&license),
-                                Color::White.bold().paint(krates.len().to_string())
-                            )?;
-                        } else {
-                            write!(output, "{license} ({}): ", krates.len())?;
-                        }
-
-                        for (i, krate_id) in krates.iter().enumerate() {
-                            if i != 0 {
-                                write!(output, ", ")?;
-                            }
-
-                            if color {
-                                let krate = crate_layout.search(krate_id);
-                                let color = if krate.licenses.len() > 1 {
-                                    Color::Yellow
-                                } else {
-                                    Color::White
-                                };
-
-                                let (name, version) = krate_id.parts();
-                                write!(output, "{}@{version}", color.paint(name))?;
-                            } else {
-                                write_pid(&mut output, krate_id)?;
-                            }
-                        }
-
-                        writeln!(output)?;
-                    }
-
-                    if !license_layout.unlicensed.is_empty() {
-                        if color {
-                            write!(
-                                output,
-                                "{} ({}): ",
-                                Color::Red.paint("Unlicensed"),
-                                Color::White
-                                    .bold()
-                                    .paint(license_layout.unlicensed.len().to_string())
-                            )?;
-                        } else {
-                            write!(output, "Unlicensed ({}): ", license_layout.unlicensed.len())?;
-                        }
-
-                        for (i, krate) in license_layout.unlicensed.iter().enumerate() {
-                            if i != 0 {
-                                write!(output, ", ")?;
-                            }
-
-                            write_pid(&mut output, krate)?;
-                        }
-
-                        writeln!(output)?;
-                    }
-                }
-                Layout::Crate => {
-                    for (id, krate) in crate_layout.crates {
-                        let (name, version) = id.parts();
-
-                        if color {
-                            let color = match krate.licenses.len() {
-                                1 => Color::White,
-                                0 => Color::Red,
-                                _ => Color::Yellow,
-                            };
-
-                            write!(
-                                output,
-                                "{}@{version} ({}): ",
-                                color.paint(name),
-                                Color::White.bold().paint(krate.licenses.len().to_string()),
-                            )?;
-                        } else {
-                            write!(output, "{name}@{version} ({}): ", krate.licenses.len(),)?;
-                        }
-
-                        for (i, license) in krate.licenses.iter().enumerate() {
-                            if i != 0 {
-                                write!(output, ", ")?;
-                            }
-
-                            if color {
-                                write!(output, "{}", Color::Cyan.paint(license))?;
-                            } else {
-                                write!(output, "{license}")?;
-                            }
-                        }
-
-                        writeln!(output)?;
-                    }
-                }
-            }
-
-            std::io::Write::write_all(&mut std::io::stdout(), output.as_bytes())?;
-        }
-        OutputFormat::Json => match args.layout {
-            Layout::License => {
-                serde_json::to_writer(std::io::stdout(), &license_layout)?;
-            }
-            Layout::Crate => serde_json::to_writer(std::io::stdout(), &crate_layout.crates)?,
-        },
-        OutputFormat::Tsv => {
-            // We ignore the layout specification and always just do a grid of crate rows x license/exception columns
-            let mut output = String::with_capacity(4 * 1024);
-
-            // Column headers
-            {
-                write!(output, "crate")?;
-
-                for (license, _) in &license_layout.licenses {
-                    write!(output, "\t{license}")?;
-                }
-
-                if !license_layout.unlicensed.is_empty() {
-                    write!(output, "\tUnlicensed")?;
-                }
-
-                writeln!(output)?;
-            }
-
-            for (id, krate) in crate_layout.crates {
-                write_pid(&mut output, &id)?;
-
-                for lic in &license_layout.licenses {
-                    if lic.1.binary_search(&id).is_ok() {
-                        write!(output, "\tX")?;
-                    } else {
-                        write!(output, "\t")?;
-                    }
-                }
-
-                if krate.licenses.is_empty() {
-                    write!(output, "\tX")?;
-                }
-
-                writeln!(output)?;
-            }
-
-            std::io::Write::write_all(&mut std::io::stdout(), output.as_bytes())?;
-        }
-    }
+    cargo_deny::licenses::list(
+        &mut std::io::stdout(),
+        &summary,
+        args.format,
+        args.layout,
+        colorize,
+    )?;
 
     Ok(())
 }
